feat: move sound_device connector to rai_s2s
